Fix #7750 shared release version parsing
