+++ crates/larch-cli/src/release_common.rs
@@ -1,0 +1,11 @@
+pub fn semver(value: &str) -> Option<(u64, u64, u64)> {
+    let mut parts = value.split('.');
+    let major = parts.next()?.parse().ok()?;
+    let minor = parts.next()?.parse().ok()?;
+    let patch = parts.next()?.parse().ok()?;
+    (parts.next().is_none()
+        && value
+            .bytes()
+            .all(|byte| byte.is_ascii_digit() || byte == b'.'))
+    .then_some((major, minor, patch))
+}
